IMERG reader bug fix - precompile commands
The read_imerghdf subroutine has precompiler instructions that require
USE_HDF5 to be defined in order for the subroutine to read the IMERG data.
These definitions are now in the LIS_misc.h file and were not included before
this fix. The user would see the prompt that the file was being read, but the subroutine
would not execute. LIS_misc.h is now added as an included file.

One of the error messages was also set to be printed to the screen. It is now
printing to the LIS log, which is similar to what occurs with the other
error statements.