IMERG LIS reader filename update
This code modification updates the LIS reader to read in version
V05B (latest IMERG algorithm version available) instead of V03D.
This code has also added a config option to allow the user to switch
between the three release versions of IMERG (early, late, and final),
all of which have different filename conventions. If no version is
declared in the configuration file, the reader defaults to the final
version.