IMERG Precipitation Reader for LVT
This update adds a reader to LVT that reads in GPM IMERG precipitation
data. This reader only works with the latest IMERG release (V05B) and
includes a configuration option that allows the user to switch between
the early, late, and final versions of the IMERG product.

Version #2: Removed excessive whitespace from code.