Fixes issue #57
The format of the TS_LOCATIONS.TXT file was changed to use
percentage of the total number of locations (instead of a minimum
number data points) as the threshold when time series data is
written out. When time series data at point locations are output,
then this percentage needs to be set to 0.0 to ensure data output.
If there are gaps in the data, then this could lead to divide by
zero errors. This has been fixed by adding additional checks
in the calculations.

Resolves #57