Added SAVE attribute for dimID in LVT_writevar_data_header.
This will preserve the netCDF dimension IDs for multiple calls to the
subroutine.  This is important, because the dimension IDs for latitude and
longitude are only set during the first subroutine call.

When Intel compilers are used, the values of dimID were preserved between
subroutine calls.  But this is not the case with GNU compilers, and the
Fortran standard only guarantees preservation if the local variable has the
SAVE attribute.