Fix LDT crash by only writing OPTUE parameters when in that run mode
This update will only write OPTUE parameter header and parameter
data when in "OPTUE parameter processing" LDT run mode.  The code
previously would crash when in "LSM parameter processing" mode.
A simple check for the runmode before calling the routines to
write the OPTUE parameters header and the data to the lis_input
netCDF file was added to the code.

Resolves: #197
See also: #183