Fix to LDT little_endian issues for GTOPO, SRTM, and TBOT
This commit fixes issues relating to configuring LDT with
little_endian, but trying to read the natively big_endian
datasets from GTOPO, SRTM, and TBOT.

The primary fixes are to the "???_Native" options for the
topography and to the "ISLCP1" option for TBOT, as these
native datasets are big_endian.

However, related fixes were also made to the "???_LIS"
options for datasets provided by the LIS team.  These
datasets are legacy data, and have traditionally been
produced in big_endian.  Users should consider moving
to the "???_Native" datasets for these parameters.

Resolves: #199
Resolves: #200