Updated RHMin calculation in NoahMP.4.0.1.
The earlier version used the CALHUM subroutine in NoahMP.4.0.1 to calculate
saturation specific humidity, following similar logic in Noah.3.6.  However,
the NoahMP version actually calculates saturation mixing ratio.  This
led to a different (and physically inconsistent) value of RHMin compared
to the values produced by Noah.3.9 and JULES.5.0 with identical atmospheric
forcing.

The new code takes the same approach as JULES.5.0, with necessary calculations
done in-line in NoahMP401_main but with formula matching the Noah.3.6 code.