Add residual snow fix option to Noah.3.6
This commit adds a fix to Noah.3.6 that will remove small amounts
of snow that should normally be cleared but are not in Noah v3.6.

This fix will zero out snow on a model point if it is below a certain
threshold. Since this fix is not an official NCAR release, the fix
is added as a config option.
The user must specify 'Noah.3.6 removal of residual snow fix: 1'
to turn the fix on. The fix is turned off (option 0) by default.

Resolves: #407