Fix LVT IMERG hard-coded filename option
The IMERG reader in LVT was still hard-coded to V05B and an old
filename extension convention. This fix will add an option for the
user to specify the version and product of IMERG data (similar to
the IMERG reader in LIS). The default options will be set to version
V06B and 'final' product.

Resolves: #180