Add max snow depth option for NoahMP 4.0.1 glacier model
This commit adds a config option that allows the user to set the
maximum snow depth (in mm) that is used in the Noah MP 4.0.1 glacier
model. This value was hard-coded to 2000. If the user does not specify
a value, the max snow depth will be set to the default value of 2000.
This change only affects NoahMP 4.0.1.

Resolves: #569