Prevent divide-by-zero in Noah-MP with frozen soils
This pull request will prevent a divide-by-zero in the
Noah-MP-4.0.1 physics when removing sub-surface runoff.
This divide-by-zero is very rare, but can occur when
the soils are totally frozen.  The sub-surface runoff
will be zero as well as the hydraulic conductivity.
A simple check of the hydraulic conductivity was added
to prevent the divide-by-zero.